Added support for file-containing-symbol reflection request.
Motivation:

The file-containing-symbol request is part of the Reflection Service and enables users to find the proto file
containing a symbol they specify and its transitive dependencies.

Modifications:

Added a dictionary of the fully qualified names of symbols and their corresponding file names, in the ReflectionServiceData struct.
Added the function that creates the server response, after getting the file name corresponding to the symbol name and
getting its tranaitive dependencies. Also, split the tests into Integration and Unit tests, and added tests for the new request.

Result:

Users of the Reflection Service will be able to get from the server the proto file that contains the symbols they are specifying in the request and its transitive dependencies.
