+++ Sources/GRPCReflectionService/Server/ReflectionService.swift
@@ -46,10 +46,13 @@ internal struct ReflectionServiceData: Sendable {
 
   internal var fileDescriptorDataByFilename: [String: FileDescriptorProtoData]
   internal var serviceNames: [String]
+  internal var fileNameBySymbol: [String: String]
 
   internal init(fileDescriptors: [Google_Protobuf_FileDescriptorProto]) throws {
     self.serviceNames = []
     self.fileDescriptorDataByFilename = [:]
+    self.fileNameBySymbol = [:]
+
     for fileDescriptorProto in fileDescriptors {
       let serializedFileDescriptorProto: Data
       do {
@@ -67,6 +70,17 @@ internal struct ReflectionServiceData: Sendable {
       )
       self.fileDescriptorDataByFilename[fileDescriptorProto.name] = protoData
       self.serviceNames.append(contentsOf: fileDescriptorProto.service.map { $0.name })
+      for qualifiedSybolName in fileDescriptorProto.qualifiedSymbolNames {
+        if let value = self.fileNameBySymbol[qualifiedSybolName] {
+          throw GRPCStatus(
+            code: .alreadyExists,
+            message:
+              "The \(qualifiedSybolName) symbol from \(fileDescriptorProto.name) already exists in \(value)."
+          )
+        } else {
+          self.fileNameBySymbol[qualifiedSybolName] = fileDescriptorProto.name
+        }
+      }
     }
   }
 
@@ -98,6 +112,16 @@ internal struct ReflectionServiceData: Sendable {
       visited.insert(currentFileName)
     }
     return serializedFileDescriptorProtos
+  }
+
+  internal func getFileNameBySymbol(named symbolName: String) throws -> String {
+    guard let fileName = self.fileNameBySymbol[symbolName] else {
+      throw GRPCStatus(
+        code: .notFound,
+        message: "The provided symbol could not be found."
+      )
+    }
+    return fileName
   }
 }
 
@@ -139,6 +163,14 @@ internal final class ReflectionServiceProvider: Reflection_ServerReflectionAsync
     )
   }
 
+  internal func findFileBySymbol(
+    _ symbolName: String,
+    request: Reflection_ServerReflectionRequest
+  ) throws -> Reflection_ServerReflectionResponse {
+    let fileName = try self.protoRegistry.getFileNameBySymbol(named: symbolName)
+    return try self.findFileByFileName(fileName, request: request)
+  }
+
   internal func serverReflectionInfo(
     requestStream: GRPCAsyncRequestStream<Reflection_ServerReflectionRequest>,
     responseStream: GRPCAsyncResponseStreamWriter<Reflection_ServerReflectionResponse>,
@@ -155,6 +187,13 @@ internal final class ReflectionServiceProvider: Reflection_ServerReflectionAsync
 
       case .listServices:
         let response = try self.getServicesNames(request: request)
+        try await responseStream.send(response)
+
+      case let .fileContainingSymbol(symbolName):
+        let response = try self.findFileBySymbol(
+          symbolName,
+          request: request
+        )
         try await responseStream.send(response)
 
       default:
@@ -185,5 +224,42 @@ extension Reflection_ServerReflectionResponse {
       $0.originalRequest = request
       $0.listServicesResponse = listServicesResponse
     }
+  }
+}
+
+extension Google_Protobuf_FileDescriptorProto {
+  var qualifiedServiceAndMethodNames: [String] {
+    var qualifiedServiceAndMethodNames: [String] = []
+
+    for service in self.service {
+      qualifiedServiceAndMethodNames
+        .append(self.package + "." + service.name)
+      qualifiedServiceAndMethodNames
+        .append(
+          contentsOf: service.method
+            .map { self.package + "." + service.name + "." + $0.name }
+        )
+    }
+    return qualifiedServiceAndMethodNames
+  }
+
+  var qualifiedMessageTypes: [String] {
+    return self.messageType.map {
+      self.package + "." + $0.name
+    }
+  }
+
+  var qualifiedEnumTypes: [String] {
+    return self.enumType.map {
+      self.package + "." + $0.name
+    }
+  }
+
+  var qualifiedSymbolNames: [String] {
+    [
+      self.qualifiedServiceAndMethodNames,
+      self.qualifiedMessageTypes,
+      self.qualifiedEnumTypes,
+    ].flatMap { $0 }
   }
 }
